scripts/requirements.txt
gcloud>=0.7.1
google-api-python-client>=1.3.2
google-apitools>=0.4.10
pyYAML>=3.11
oauth2client>=1.4.12
httplib2==0.9.1
